Prioritize
This function prioritize accepts an array and a callback. The callback returns either true or false. prioritize will iterate through the array and perform the callback on each element, and return a new array, where all the elements that yielded a return value of true come first in the array, and the rest of the elements come second.